+++ src/main/java/com/tackpad/dao/MessageDao.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface MessageDao extends BaseDao<Message> {
 	List<Message> getPage(int page, int pageSize, List<Long> messageIdList, Long companyBranchId, Long companyId, List<Long> companyCategoryIdList,
 						  List<Long> companyCategoryMainIdList, List<MessageType> messageTypeList, Double latitude,
-						  Double longitude, Integer range, String searchTerm, ListingSortType listingSortType) throws ParseException;
+						  Double longitude, Double range, String searchTerm, ListingSortType listingSortType) throws ParseException;
 
     Long findCount(Long companyId);
 }
